Don’t use SchemaMonkey module in spec
It’s neat & clever to define a SchemaMonkey client to add support methods.  But I think we shouldn’t:  We’re testing additions & modifications to AR, we shouldn’t introduce additional ones as part of the testing.

Also, while the helper was required and registered in this file, the effect is actually global, which could be confusing.